Underground update
*Implements underground: each team has an underground map. The king ant can dig through.
*Implements nest. A nest is generated at a random place at map generation. The king ant can only dig new exit from the nest's underground. The new exit is placed at a random place on the map.
*Changed coordinates representation from half-assed tuple to dedicated class

Bug fix:
*clicking outside of the map doesn't crash the game anymore

Notes:
I settled with giving the first underground tile row's a Z index of 1. It means it's possible to know if a creature is outside or underground by looking at its coordinate only, which is good. It also means that the underground tile map index is different from the creatures' Z index, which will cause problem if forgotten